third_party/blink/web_tests/virtual/webnn-service-on-cpu/external/wpt/webnn/conformance_tests/pooling.https.any_cpu-expected.txt

This is a testharness.js-based test.
[FAIL] averagePool2d float32 4D tensor options.padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 41.9493293762207 should be close enough to expected 52.43666076660156 by ULP distance: expected a number less than or equal to 27n but got 2749191n
[FAIL] averagePool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] averagePool2d float32 4D tensor options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 36.13502502441406 should be close enough to expected 54.20252990722656 by ULP distance: expected a number less than or equal to 11n but got 4736288n
[FAIL] averagePool2d float32 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 36.13502502441406 should be close enough to expected 54.20252990722656 by ULP distance: expected a number less than or equal to 11n but got 4736288n
[FAIL] averagePool2d float32 4D tensor options.roundingType=ceil and no padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 50.63130569458008 should be close enough to expected 40.29140853881836 by ULP distance: expected a number less than or equal to 11n but got 2710542n
[FAIL] averagePool2d float32 4D tensor options.layout=nhwc and options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 36.13502502441406 should be close enough to expected 54.20252990722656 by ULP distance: expected a number less than or equal to 11n but got 4736288n
[FAIL] averagePool2d float32 4D tensor options.layout=nhwc and options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 36.13502502441406 should be close enough to expected 54.20252990722656 by ULP distance: expected a number less than or equal to 11n but got 4736288n
[FAIL] averagePool2d float32 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 36.13502502441406 should be close enough to expected 54.20252990722656 by ULP distance: expected a number less than or equal to 11n but got 4736288n
[FAIL] averagePool2d float32 4D tensor options.outputSizes ignores options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 36.13502502441406 should be close enough to expected 54.20252990722656 by ULP distance: expected a number less than or equal to 11n but got 4736288n
[FAIL] averagePool2d float32 4D tensor options.dilations with options.strides
  assert_less_than_equal: assert_array_approx_equals_ulp: test averagePool2d float32 actual 28.626827239990234 should be close enough to expected 42.940242767333984 by ULP distance: expected a number less than or equal to 11n but got 4636433n
[FAIL] l2Pool2d float32 4D constant tensor all positive default options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor default all positive options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor default all negative options
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.windowDimensions
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.padding
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.strides
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.layout=nchw
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.layout=nhwc
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.roundingType=floor
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.roundingType=ceil
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.outputSizes ignores options.roundingType=floor
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.outputSizes ignores options.roundingType=ceil
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] l2Pool2d float32 4D tensor options.dilations with options.strides
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type float32 must be one of []."
[FAIL] maxPool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] maxPool2d float32 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual 99.28312683105469 should be close enough to expected -39.03501892089844 by ULP distance: expected a number less than or equal to 0n but got 2229449938n
[FAIL] maxPool2d float32 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual 99.28312683105469 should be close enough to expected -39.03501892089844 by ULP distance: expected a number less than or equal to 0n but got 2229449938n
Harness: the test ran to completion.
